Generic method for sorting and pagination
