+++ controllers/bootcamps.js
@@ -10,71 +10,7 @@ const Bootcamp = require('../models/Bootcamp');
 
 exports.getBootcamps = asyncHandler( async (req, res, next)=>{
 
-    let query;
-
-    // Make a copy of the query
-    const reqQuery = {...req.query};
-    // Remove fields from query
-    const removeFields = ['select', 'sort', 'limit', 'page'];
-
-    // Loop over removeFields  and detele from query
-    removeFields.forEach(param => delete reqQuery[param]);
-    // Create query
-    let queryStr = JSON.stringify(reqQuery);
-    // console.log(queryStr);
-    // Create logical operators ie gte, lt, lte, etc
-    queryStr = queryStr.replace(/\b(gt|gte|lt|lte|in)\b/g, match =>`$${match}`);
-    // Finding resource
-    query = Bootcamp.find(JSON.parse(queryStr)).populate('courses');
-    
-
-    // Select Fields
-    if (req.query.select) {
-        const fields = req.query.select.split(',').join(' ');
-        query = query.select(fields);
-    }
-
-    // Sorting the fields
-    if (req.query.sort) {
-        const sortBy = req.query.sort.split(',').join(' ');
-        query = query.sort(sortBy);
-        console.log(sortBy);
-    }
-    else{
-        query = query.sort('-createdAt');
-    }
-
-    // Pagination
-    const page = parseInt(req.query.page, 10)|| 1;
-    const limit = parseInt(req.query.limit, 10)|| 1;
-    const startIndex = (page-1)*limit;
-    const endIndex = page*limit;
-    const total = await Bootcamp.countDocuments();
-
-    query = query.skip(startIndex).limit(limit);
-
-
-
-    // Executing query
-        const bootcamps = await query;
-
-        // Pagination
-        const pagination = {};
-        if (endIndex < total) {
-            pagination.next = {
-                page: page + 1,
-                limit:limit
-            };
-        }
-
-        if (startIndex > 0) {
-            pagination.prev ={
-                page: page - 1,
-                limit:limit
-            };
-        }
-        res.status(200).json({success: true, count: bootcamps.length, pagination, data: bootcamps});
-        
+        res.status(200).json(res.advancedResults);
 
 });
