+++ controllers/courses.js
@@ -9,24 +9,20 @@ const Bootcamp = require('../models/Bootcamp');
 // @access  Public
 
 exports.getCourses = asyncHandler (async (req, res, next)=>{
-    let query;
+
 
     if (req.params.bootcampId) {
-       query = Course.find({ bootcamp: req.params.bootcampId});
-    }
-    else{
-        query = Course.find().populate({
-            path: 'bootcamp',
-            select: 'name description'
-        });
-    }
-    const courses = await query;
-    
-    res.status(200).json({
+       const courses = await Course.find({ bootcamp: req.params.bootcampId});
+       return res.status(200).json({
         success: true,
         count: courses.length, 
         data: courses
-    });
+       });
+    }
+    else{
+       res.status(200).json(res.advancedResults);
+    }
+   
 });
 
 // @desc    Get Course
