+++ routes/bootcamps.js
@@ -8,6 +8,9 @@ const {
     deleteBootcamp,
     getBootcampsWithinRadius, 
     uploadBootcampPhoto} = require('../controllers/bootcamps');
+
+const advancedResults = require('../middleware/advancedRequest');
+const Bootcamp = require('../models/Bootcamp');
 // Include routers for other resources
 const courseRouter = require('./courses');
 
@@ -20,7 +23,8 @@ router.route('/radius/:zipcode/:distance').get(getBootcampsWithinRadius);
 router.route('/:id/photo').put(uploadBootcampPhoto);
 
 
-router.route('/').get(getBootcamps).post(createBootcamp);
+router.route('/').get(advancedResults(Bootcamp, 'courses'),
+ getBootcamps).post(createBootcamp);
 router.route('/:id').get(getBootcamp).put(updateBootcamp).delete(deleteBootcamp);
 
 module.exports = router;
